font-awesome/requirements.txt
requests==2.12.1
PyYAML==3.12
